+++ src/functions_test.py
@@ -32,7 +32,7 @@ def make_test_units():
     red_id = 0
     for i in range(15):
         if red_id == 6 or red_id == 7: LIST_WITH_UNITS.append(Main_battle_tank([300 + 100*i, 1500], -math.pi/2, 2, 2)) # 2500      
-        else: LIST_WITH_UNITS.append(Light_tank([300 + 100*i, 1000], -math.pi/2, 2, 2))
+        else: LIST_WITH_UNITS.append(Light_tank([300 + 100*i, 1500], -math.pi/2, 2, 2))
         LIST_WITH_UNITS[id].base.movement_target = [[500 + 50*i, 400]]
         id += 1
         red_id += 1